packages/tmux_manager/pyproject.toml
[project]
name = "ai-tmux-manager"
version = "0.1.0"
description = "Tmux session manager for AI Infrastructure agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-core",
    "libtmux>=0.28.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_tmux"]
